requirements.txt
python-dotenv==1.0.0
boto3==1.29.0